+++ src/main/java/com/weiketang/uclassrear/service/MyProjectServiceImp.java
@@ -24,31 +24,37 @@ public class MyProjectServiceImp implements MyProjectService{
 
     @Override
     public MyProject createOneProject(String publisherId, MyProject myProject) {
+        if(publisherId == null || myProject == null) return null;
         return (MyProject)myModelDao.createOneModel(publisherId, myProject, new MyProject().getModelType());
     }
 
     @Override
     public DeleteResult removeOneProjectById(String projectId) {
+        if(projectId == null) return null;
         return myModelDao.removeOneModelById(projectId, new MyProject().getModelType());
     }
 
     @Override
     public ModelJoinRecord joinOneProject(String participatorId, String projectId) {
+        if(participatorId == null || projectId == null) return null;
         return myModelDao.joinOneModel(participatorId, projectId);
     }
 
     @Override
-    public DeleteResult exitOneCourse(String participatorId, String projectId) {
+    public DeleteResult exitOneProject(String participatorId, String projectId) {
+        if(participatorId == null || projectId == null) return null;
         return myModelDao.exitOneModel(participatorId, projectId);
     }
 
     @Override
     public List<MyModel> getProjectsByPublisher(String publisherId) {
-        return myModelDao.getModelsByPublisher(publisherId);
+        if(publisherId == null) return null;
+        return myModelDao.getModelsByPublisher(publisherId, new MyProject().getModelType());
     }
 
     @Override
     public List<MyModel> getProjectsByParticipator(String participatorId) {
+        if(participatorId == null) return null;
         return myModelDao.getModelsByParticipator(participatorId, new MyProject().getModelType());
     }
 
@@ -59,26 +65,31 @@ public class MyProjectServiceImp implements MyProjectService{
 
     @Override
     public List<User> getParticipatorsByProject(String projectId) {
+        if(projectId == null) return null;
         return myModelDao.getParticipatorsByModel(projectId);
     }
 
     @Override
     public List<MyModel> searchProjects(String inputStr) {
+        if(inputStr == null) return null;
         return myModelDao.searchModels(inputStr, new MyProject().getModelType());
     }
 
     @Override
     public MyFavorModel collectOneProject(String projectId, String userId) {
+        if(projectId == null || userId == null) return null;
         return myModelDao.collectOneModel(projectId, userId);
     }
 
     @Override
     public List<MyFavorModel> getFavorProjectsByUserId(String userId) {
+        if(userId == null) return null;
         return myModelDao.getFavorModelsByUserId(userId, new MyProject().getModelType());
     }
 
     @Override
     public DeleteResult removeOneFavorProject(String projectId, String userId) {
+        if(projectId == null || userId == null) return null;
         return myModelDao.removeOneFavorModel(projectId, userId);
     }
 }
